enforce Universe.dimensions.dtype is always np.float32

+++ package/MDAnalysis/coordinates/memory.py
@@ -372,10 +372,10 @@ class MemoryReader(base.ProtoReader):
         self.ts = self._Timestep(self.n_atoms, **kwargs)
         self.ts.dt = dt
         if dimensions is None:
-            dimensions = np.zeros((self.n_frames, 6), dtype=np.float64)
+            dimensions = np.zeros((self.n_frames, 6), dtype=np.float32)
         else:
             try:
-                dimensions = np.asarray(dimensions, dtype=np.float64)
+                dimensions = np.asarray(dimensions, dtype=np.float32)
             except ValueError:
                 raise TypeError("'dimensions' must be array-like got {}"
                                 "".format(type(dimensions)))
